separate cache from node
