--- apps/neoscan_monitor/lib/neoscan_monitor.ex
@@ -1,18 +1,0 @@
-defmodule NeoscanMonitor do
-  @moduledoc """
-  Documentation for NeoscanMonitor.
-  """
-
-  @doc """
-  Hello world.
-
-  ## Examples
-
-      iex> NeoscanMonitor.hello
-      :world
-
-  """
-  def hello do
-    :world
-  end
-end
