+++ apps/neoscan_monitor/lib/neoscan_monitor/monitor/utils.ex
@@ -5,87 +5,6 @@ defmodule NeoscanMonitor.Utils do
   alias Neoscan.Stats
   alias Neoscan.ChainAssets
   alias NeoscanNode.Notifications
-  alias NeoscanNode.Blockchain
-
-  # blockchain api nodes
-  def seeds do
-    Application.fetch_env!(:neoscan_monitor, :seeds)
-  end
-
-  # function to load nodes state
-  def load do
-    data =
-      seeds()
-      |> pmap(fn url ->
-        current_height = Blockchain.get_current_height(url)
-        {url, current_height, evaluate_result(url, current_height)}
-      end)
-      |> Enum.filter(fn {_, _, keep} -> keep end)
-      |> Enum.map(fn {url, {:ok, height}, _} -> {url, height} end)
-
-    set_state(data)
-  end
-
-  def pmap(collection, func) do
-    collection
-    |> Enum.map(&Task.async(fn -> func.(&1) end))
-    |> Enum.map(&Task.await(&1, 15_000))
-  end
-
-  # handler for nil data
-  defp set_state([] = data) do
-    %{:nodes => [], :height => {:ok, nil}, :data => data}
-  end
-
-  # call filters on results and set state
-  defp set_state(data) do
-    height = filter_height(data)
-    %{nodes: filter_nodes(data, height), height: {:ok, height}, data: data}
-  end
-
-  # filter working nodes
-  defp filter_nodes(data, height) do
-    data
-    |> Enum.filter(fn {_url, hgt} -> hgt == height end)
-    |> Enum.map(fn {url, _height} -> url end)
-  end
-
-  # filter current height
-  defp filter_height(data) do
-    {height, _count} =
-      data
-      |> Enum.map(fn {_url, height} -> height end)
-      |> Enum.reduce(%{}, fn height, acc ->
-        Map.update(acc, height, 1, &(&1 + 1))
-      end)
-      |> Enum.max_by(fn {_height, count} -> count end)
-
-    height
-  end
-
-  # handler to filter errors
-  defp evaluate_result(url, {:ok, height}) do
-    test_get_block(url, height)
-  end
-
-  defp evaluate_result(_url, {:error, _height}) do
-    false
-  end
-
-  # test node api
-  defp test_get_block(url, height) do
-    Blockchain.get_block_by_height(url, height - 1)
-    |> test()
-  end
-
-  # handler to test response
-  defp test({:ok, _block}) do
-    true
-  end
-
-  defp test({:error, _reason}) do
-    false
-  end
 
   # function to cut extra elements
   def cut_if_more(list, count) when count == 15 do
