+++ apps/neoscan_monitor/test/neoscan_monitor_test.exs
@@ -1,4 +1,3 @@
 defmodule NeoscanMonitorTest do
   use ExUnit.Case
-  doctest NeoscanMonitor
 end
