+++ apps/neoscan_node/lib/neoscan_node.ex
@@ -1,3 +1,12 @@
 defmodule NeoscanNode do
   @moduledoc false
+  alias NeoscanNode.Worker
+
+  def get_nodes do
+    Worker.get_nodes()
+  end
+
+  def get_height do
+    Worker.get_height()
+  end
 end
