+++ apps/neoscan_sync/lib/neoscan_sync/genstage/producer.ex
@@ -5,7 +5,6 @@ defmodule NeoscanSync.Producer do
   alias NeoscanNode.Blockchain
   alias NeoscanNode.Notifications
   alias NeoscanNode.HttpCalls
-  alias NeoscanMonitor.Api, as: MonitorApi
   alias Neoscan.Blocks
 
   require Logger
@@ -83,7 +82,7 @@ defmodule NeoscanSync.Producer do
   end
 
   defp check_if_nodes(n) do
-    nodes = HttpCalls.url(MonitorApi.get_nodes(), n)
+    nodes = HttpCalls.get_url(n)
 
     if Enum.count(nodes) == n do
       nodes
@@ -118,7 +117,7 @@ defmodule NeoscanSync.Producer do
 
   # get current height from monitor
   def get_current_height do
-    MonitorApi.get_height()
+    NeoscanNode.get_height()
   end
 
   def add_notifications(block, height) do
